refactor: renamed folder Fields and updated import paths for field components
