+++ src/app/auth/Auth.tsx
@@ -6,7 +6,7 @@ import { useForm } from 'react-hook-form'
 import { Logo } from '@/components/layout/sidebar/header/Logo'
 import { SkeletonLoader } from '@/ui/SkeletonLoader'
 import { Button } from '@/ui/button/Button'
-import { Field } from '@/ui/field/Field'
+import { Field } from '@/ui/fields/Field'
 import { useAuthForm } from './useAuthForm'
 import type { IAuthForm } from '@/types/auth-form.types'
 import styles from './captcha.module.scss'
